pyproject.toml
[project]
name = "transcribe-audio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deep-translator>=1.11.4",
    "openai-whisper>=20250625",
    "pyside6>=6.11.0",
    "rich>=15.0.0",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.26.0",
]

[dependency-groups]
dev = [
    "commitizen>=4.13.10",
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "1.5.0"
version_files = ["pyproject.toml:version"]
tag_format = "v$version"
update_changelog_on_bump = true
bump_message = "bump: version $current_version → $new_version [skip ci]"
